+++ youlink/src/views/Home.vue
@@ -23,31 +23,7 @@ YOULINKで募集させていただきました。</p>
 <script>
 import HomeModal from '../components/HomeModal.vue'
 import { ref, onMounted } from 'vue'
-import { collection, getDocs } from 'firebase/firestore'
-import { db } from '../firebaseConfig'
 
-const data = ref()
-const isLoading = ref(false)
-const error = ref()
-
-const getUsers = async () => {
-    console.log("getUsers")
-    isLoading.value = true
-    try {
-        const usersCollection = collection(db, 'users');
-        const usersSnapshot = await getDocs(usersCollection);
-        data.value = usersSnapshot.docs.map(doc => doc.data());
-        console.log(data.value)
-    } catch (e) {
-        error.value = e
-    } finally {
-        isLoading.value = false
-    }
-}
-
-onMounted(() => {
-    getUsers()
-})
 export default ({
   components: {
     HomeModal,
