+++ youlink/src/views/Top.vue
@@ -1,0 +1,8 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+    <RouterLink to="/signin"><button>SignIn</button></RouterLink>
+    <RouterLink to="/signup"><button>SignUp</button></RouterLink>
+</template>
